requirements.txt
nltk==3.8.1
# LangChain Core
langchain
langchain-core

# Google Gemini Integration
langchain-google-genai
google-generativeai

# Environment Variable Management
python-dotenv

# Machine Learning Utilities
numpy
scikit-learn

# other dependencies
langchain_experimental
langchain-community
langgraph